Refactored app to match common Flask practices.
