+++ .gitignore
@@ -1,2 +1,2 @@
-config.py
+Dashboard/config.py
 Data/
